+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Refactoring in habit definition page: move logic to cubit
+
+## [0.8.199] - 2022-11-20
 ### Fixed:
-- Show expected habit success when defined
+- Show expected habit success from when defined
 
 ## [0.8.198] - 2022-11-20
 ### Added: